Moved the resolve utility function to component test support
