+++ test/component/cypress/support/utils.js
@@ -1,0 +1,2 @@
+export const resolve = (options) =>
+  new Promise((resolve) => setTimeout(resolve, options.after, options.with))
