+++ test/sandbox/utils.js
@@ -4,7 +4,7 @@
  * `faker` and `json-schema-faker` _seeded_ with a constant _seed_.
  *
  * The seed configuration only applies to the currently imported instance of
- * each library so we only should instances exported from this module.
+ * each library so we only should use instances exported from this module.
  *
  * We enforce this with the `no-restricted-imports` ESLint rule and this should
  * be the only valid place to violate that rule.
@@ -18,9 +18,6 @@ import seedrandom from 'seedrandom'
 const SEED = 123
 
 fkr.seed(SEED)
-
-export const resolve = (options) =>
-  new Promise((resolve) => setTimeout(resolve, options.after, options.with))
 
 JSONSchemaFaker.option({
   random: seedrandom(SEED),
